# back-end/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The official documentation of StudyTogether",
    "version": "1.0.5",
    "title": "StudyTogether's Offical API Documentation"
  },
...
  "paths": {
...
    "/studygroups": {
      "get": {
        "tags": ["Study Groups"],
        "summary": "Returns all the study groups in the database",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about all study groups"
              }
            }
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/saved": {
      "get": {
        "tags": ["Study Groups"],
        "summary": "Returns all the study groups the current logged in user has bookmarked",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about all study groups"
              }
            }
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/registered": {
      "get": {
        "tags": ["Study Groups"],
        "summary": "Returns all the study groups the current logged in user has registered to attend",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about all study groups"
              }
            }
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/{id}": {
      "get": {
        "tags": ["Study Groups"],
        "summary": "Find study group by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of study group to return",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about a specific study group"
              }
            }
          },
          "400": {
            "description": "study group with the given id was not found in database"
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/attend/{id}": {
      "post": {
        "tags": ["Study Groups"],
        "summary": "Attend a study group as a logged in user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of study group to attend",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about a specific study group"
              }
            }
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "404": {
            "description": "Study group with the given id was not found in database"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/create": {
      "post": {
        "tags": ["Study Groups"],
        "summary": "Create a study group",
        "description": "The user must be authenticated since their user id will be used as the study group's host id",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "body",
            "description": "Title of study group",
            "required": true,
            "type": "string"
          },
          {
            "name": "startDateTime",
            "in": "body",
            "description": "The time when the study group meeting starts",
            "required": true,
            "type": "date"
          },
          {
            "name": "endDateTime",
            "in": "body",
            "description": "The time when the study group meeting ends",
            "required": true,
            "type": "date"
          },
          {
            "name": "phone",
            "in": "body",
            "description": "The phone number associated with the study group",
            "required": true,
            "type": "string"
          },
          {
            "name": "imgUrl",
            "in": "body",
            "description": "The study group's image",
            "required": true,
            "type": "string"
          },
          {
            "name": "curAttendees",
            "in": "body",
            "description": "The current number of attendees of the study group",
            "default": 0,
            "minimum": 0,
            "required": false,
            "type": "integer"
          },
          {
            "name": "location",
            "in": "body",
            "description": "The coordinates (longitude and latitude) of the meeting",
            "required": true,
            "type": "{long: x, lat: y}"
          },
          {
            "name": "maxAttendees",
            "in": "body",
            "description": "The maximum number of attendees of the study group",
            "required": true,
            "type": "integer"
          },
          {
            "name": "hostId",
            "in": "body",
            "description": "The id of the study group host",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "body",
            "description": "The study group description",
            "required": true,
            "type": "string"
          },
          {
            "name": "official",
            "in": "body",
            "description": "whether the study group is an offical office hours session",
            "required": true,
            "type": "bool"
          },
          {
            "name": "tags",
            "in": "body",
            "description": "The study group's tags",
            "required": true,
            "type": "[string]"
          },
          {
            "name": "recurring",
            "in": "body",
            "description": "The frequency at which this study group will recur",
            "required": true,
            "type": "string"
          },
          {
            "name": "endDateTime",
            "in": "body",
            "description": "The time when the study group meeting ends",
            "required": true,
            "type": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about a specific study group"
              }
            }
          },
          "400": {
            "description": "study group with the given id was not found in database"
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/studygroups/leave/{id}": {
      "patch": {
        "tags": ["Study Groups"],
        "summary": "Leave a study group that the current logged in user has registered to attend",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of study group to leave",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about a specific study group"
              }
            }
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "404": {
            "description": "Study group with the given id was not found in database"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/studygroups/delete/{id}": {
      "delete": {
        "tags": ["Study Groups"],
        "summary": "Delete a study group",
        "description": "A user can only delete study groups they have created",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of study group to delete",
            "required": true,
            "type": "string"
          },
          {
            "name": "deleteAll",
            "in": "body",
            "description": "Designate whether to delete this single meeting or all recurring meetings",
            "required": true,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "object containing information about a specific study group"
              }
            }
          },
          "400": {
            "description": "study group with the given id was not found in database"
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}